microservice_interface/requirements.txt
aiormq==6.4.2
anyio==3.6.2
asttokens==2.0.5
carrot-rpc==0.2.8
click==8.1.3
executing==0.8.1
fastapi==0.85.1
h11==0.14.0
idna==3.4
multidict==6.0.2
orjson==3.8.0
pamqp==3.2.1
pydantic==1.10.2
python-multipart==0.0.5
simple-print==1.4.7
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
termcolor==1.1.0
typing_extensions==4.4.0
uvicorn==0.18.3
yarl==1.8.1
